site/requirements.txt
asgiref==3.5.2
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.0
cryptography==37.0.4
defusedxml==0.7.1
Deprecated==1.2.13
distro==1.7.0
Django==3.2.14
django-allauth==0.51.0
django-braces==1.15.0
django-cors-headers==3.13.0
django-rest-framework==0.1.0
djangorestframework==3.13.1
idna==3.3
loguru==0.6.0
oauthlib==3.2.0
psycopg2-binary==2.9.3
pycparser==2.21
pydantic==1.9.1
PyJWT==2.4.0
python-decouple==3.6
python-social-auth==0.3.6
python3-openid==3.2.0
pytz==2022.1
requests==2.28.1
requests-oauthlib==1.3.1
social-auth-app-django==5.0.0
social-auth-core==4.3.0
sqlparse==0.4.2
typing_extensions==4.3.0
urllib3==1.26.11
uuid==1.30
wrapt==1.14.1
yandex-checkout==1.6.5
